requirements.txt
pydantic
langchain-core
dotenv
langchain-openai
lightrag-hku